requirements.txt
bokeh
matplotlib
mpld3
numpy
Pillow
pyyaml
selenium
streamlit
